Only update local data if provider is present (#229)
When setting edges or mode, currently updateLocalData will be called before invalidateSafeAreaInsets meaning that _currentSafeAreaInsets will be zero. This in turn causes the shadow view to be updated with a padding (or margin) of zero. This in turn sometimes causes a 1 frame flicker (only reproducible on real device, not sim) where the safe area view is zero before being updated with the correct one.