Refactor style prop in SafeAreaContext component to avoid unnecessary re-renders.
In the `SafeAreaContext.tsx` component, the `style` prop was passed using array syntax, which created a new array on each render and caused unnecessary re-renders.

To improve performance, the `style` prop has been refactored to use the spread syntax, which merges the component's default styles with the passed styles object more efficiently. This change ensures that the `SafeAreaContext` component maintains its appearance and behavior while reducing the chance of unnecessary re-renders. The new syntax is `style={{...styles.fill, ...style}}`.